api/requirements.txt
# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis[hiredis]==5.0.1

# Kafka
confluent-kafka==2.3.0

# Pydantic & Settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client
aiohttp==3.9.1
httpx==0.26.0

# Rate Limiting
slowapi==0.1.9

# Utilities
python-multipart==0.0.7
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0

# Monitoring
prometheus-client==0.19.0
